requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
APScheduler==3.10.4
gspread==6.1.4
google-auth==2.34.0
